+++ sdk/face/azure-ai-vision-face/CHANGELOG.md
@@ -10,6 +10,14 @@
 
 ### Other Changes
 
+## 1.2.0
+- Change the default service API version to `v1.2`
+- Liveness API change
+  - Unified the `LivenessSession` and removed `CreateLivenessSessionResult`
+  - Unified the `CreateLivenessWithVerifySession`
+  - Updated model `LivenessSession` and `CreateLivenesSessionContent`
+
+
 ## 1.0.0-beta.2 (2024-10-23)
 
 - Added support for the Large Face List and Large Person Group: